+++ packages/dnb-eufemia/src/extensions/forms/hooks/index.ts
@@ -1,3 +1,9 @@
-export { default as useDataValue } from './useDataValue'
+export { default as useFieldProps } from './useFieldProps'
 export { default as useErrorMessage } from './useErrorMessage'
 export { default as useProcessManager } from './useProcessManager'
+
+/**
+ * Can be removed in v11
+ * @deprecated Use `useFieldProps` instead
+ */
+export { default as useDataValue } from './useFieldProps'
